rework of file structure in app code. renames some components, and also progress on readme for app
