+++ winkduino-app/Pages/Home/CreateCustomCommand.tsx
@@ -1,9 +1,9 @@
 import { Pressable, ScrollView, Text, View } from "react-native";
-import { useColorTheme } from "../hooks/useColorTheme";
+import { useColorTheme } from "../../hooks/useColorTheme";
 import IonIcons from "@expo/vector-icons/Ionicons";
 import { useNavigation, useRoute } from "@react-navigation/native";
 import { useEffect, useState } from "react";
-import { useBLE } from "../hooks/useBLE";
+import { useBLE } from "../../hooks/useBLE";
 import { ActivityIndicator } from "react-native";
 
 export function CreateCustomCommand() {
