+++ winkduino-app/eas.json
@@ -2,16 +2,14 @@
   "build": {
     "development": {
       "developmentClient": true,
-      "distribution": "internal"
+      "distribution": "internal",
+      "ios": {
+        "simulator": true
+      }
     },
     "preview": {
       "distribution": "internal"
     },
-    "production": {},
-    "simulator": {
-      "ios": {
-        "simulator": true
-      }
-    }
+    "production": {}
   }
 }
