+++ winkduino-app/package.json
@@ -32,7 +32,6 @@
     "react-native-screens": "~4.1.0",
     "react-native-sticky-range-slider": "^0.1.7",
     "react-native-walkthrough-tooltip": "^1.6.0",
-    "react-native-wheel-color-picker": "^1.3.1",
     "react-native-wifi-reborn": "^4.13.0",
     "rn-vertical-slider-matyno": "^1.0.1",
     "toggle-switch-react-native": "^3.3.0"
